[refactor/#87] 템플릿/작성 버튼탭 분리
